+++ sqlg-test/src/main/java/org/umlg/sqlg/test/schema/TestMultipleThreadMultipleJvm.java
@@ -41,7 +41,8 @@ public class TestMultipleThreadMultipleJvm extends BaseTest {
             configuration = configs.properties(sqlProperties);
             Assume.assumeTrue(isPostgres());
             configuration.addProperty("distributed", true);
-//            configuration.addProperty("maxPoolSize", 3);
+            configuration.addProperty("c3p0.maxPoolSize", 3);
+            configuration.addProperty("dataSource.maximumPoolSize", 3);
             if (!configuration.containsKey("jdbc.url"))
                 throw new IllegalArgumentException(String.format("SqlGraph configuration requires that the %s be set", "jdbc.url"));